requirements.txt
Pillow
numpy
opencv-python-headless<=4.5.4.60